requirements.txt
ensta==5.2.2
praw==7.7.1
pillow==10.2.0
requests==2.31.0
flask
